+++ src/client.js
@@ -129,6 +129,17 @@ const Client = function Client(config = {}) {
       });
     };
 
+    // Allow alias only for users
+    // TODO move to new hull-client-node
+    if (config.userClaim || config.accessToken) {
+      this.alias = (body) => {
+        return batch({
+          type: "alias",
+          body
+        });
+      };
+    }
+
     if (config.userClaim) {
       this.account = (accountClaim = {}) => {
         if (!accountClaim) {